fix : 소설 좋아요 수정

+++ src/main/java/com/a509/service_novel/mogo/sequencce/NovelLikeSequenceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NovelLikeSequenceRepository extends MongoRepository<NovelLikeSequence, Integer> {
+public interface NovelLikeSequenceRepository extends MongoRepository<NovelLikeSequence, Long> {
 	NovelLikeSequence findSequenceById(Integer id);
 }
